src/backend/requirements.txt
alabaster==1.0.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.4.0
docutils==0.21.2
fake-useragent==2.2.0
free_proxy==1.1.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==25.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
roman-numerals-py==3.1.0
scholarly==1.7.11
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
whitenoise==6.9.0
wrapt==1.17.2
wsproto==1.2.0
